Looking for perf !
Basically, aim costs a constant time and add 20-30% cpu (corei7).
Pathfinding and update edge take from 16 to 37% cpu, that is a way to much.
In fact, update edge doesnt costs a lot, but looking for the farest point in sight is too much to handle each tick.
So, either do not check it each tick, or find a way to reduce the cost of this operation, with precomputation for example.